switching to esbuild in gulpfile
