+++ package.json
@@ -51,7 +51,6 @@
     "build": "gulp build",
     "compile": "gulp ext:compile",
     "watch": "gulp watch",
-    "postinstall": "cd mssql-react-app && yarn install",
     "lint": "gulp lint",
     "escompile": "npm run check-types && node esbuild.js",
     "escheck-types": "tsc --noEmit",
@@ -73,6 +72,7 @@
     "@angular/upgrade": "~2.1.2",
     "@eslint/compat": "^1.1.0",
     "@eslint/js": "^9.5.0",
+    "@types/adm-zip": "^0.5.5",
     "@types/azdata": "^1.44.0",
     "@types/chai": "^4.3.16",
     "@types/ejs": "^3.1.0",
@@ -84,10 +84,14 @@
     "@types/mocha": "^5.2.7",
     "@types/node": "^18.11.9",
     "@types/node-fetch": "^2.6.2",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-resizable": "^3.0.8",
     "@types/sinon": "^10.0.12",
     "@types/tmp": "0.0.28",
     "@types/underscore": "1.8.3",
     "@types/vscode": "1.78.1",
+    "@types/vscode-webview": "^1.57.5",
     "@vscode/test-electron": "^2.3.9",
     "@xmldom/xmldom": "0.8.4",
     "angular-in-memory-web-api": "0.1.13",
@@ -104,6 +108,8 @@
     "eslint-plugin-deprecation": "^3.0.0",
     "eslint-plugin-jsdoc": "^48.2.12",
     "eslint-plugin-notice": "^1.0.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.7",
     "gulp": "^4.0.2",
     "gulp-concat": "^2.6.0",
     "gulp-eslint-new": "^2.1.0",
@@ -137,13 +143,15 @@
     "@azure/arm-subscriptions": "^5.0.0",
     "@azure/msal-common": "^11.0.0",
     "@azure/msal-node": "^1.16.0",
+    "@fluentui/react-components": "^9.54.3",
     "@microsoft/ads-extension-telemetry": "^3.0.2",
+    "@monaco-editor/react": "^4.6.0",
+    "@vscode/webview-ui-toolkit": "^1.4.0",
+    "adm-zip": "^0.5.14",
     "axios": "^1.6.0",
     "core-js": "^2.4.1",
-    "decompress-zip": "^0.2.2",
     "ejs": "^3.1.7",
     "error-ex": "^1.3.0",
-    "esbuild-plugin-clean": "^1.0.1",
     "figures": "^1.4.0",
     "getmac": "1.2.1",
     "jquery": "^3.4.1",
@@ -154,6 +162,11 @@
     "pretty-data": "^0.40.0",
     "qs": "^6.9.1",
     "rangy": "^1.3.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-markdown": "^9.0.1",
+    "react-resizable": "^3.0.5",
+    "react-router-dom": "^6.24.1",
     "reflect-metadata": "0.1.12",
     "semver": "https://registry.npmjs.org/semver/-/semver-5.0.3.tgz",
     "tar": "^6.1.9",
