+++ package.nls.json
@@ -139,6 +139,5 @@
 "mssql.objectExplorer.groupBySchema":"When enabled, the database objects in Object Explorer will be categorized by schema.",
 "mssql.objectExplorer.enableGroupBySchema":"Enable Group By Schema",
 "mssql.objectExplorer.disableGroupBySchema":"Disable Group By Schema",
-"mssql.objectExplorer.expandTimeout":"The timeout in seconds for expanding a node in Object Explorer. The default value is 45 seconds.",
-"mssql.enablePreviewFeaturesDescription": "Enable preview features in the extension. Requires a restart of VS Code to take effect."
+"mssql.objectExplorer.expandTimeout":"The timeout in seconds for expanding a node in Object Explorer. The default value is 45 seconds."
 }
