+++ src/controllers/reactWebviewController.ts
@@ -11,18 +11,15 @@ export class ReactWebViewPanelController<T> implements vscode.Disposable {
 	private _isDisposed: boolean = false;
 	private _state: T;
 	private _webViewRequestHandlers: { [key: string]: (params: any) => any } = {};
-	private _assetUri: vscode.Uri;
 	private _reducers: { [key: string]: (state: T, payload: any) => ReducerResponse<T> } = {};
 
 	constructor(
 		private _context: vscode.ExtensionContext,
 		title: string,
 		private _srcFile: string,
-		// private _styleFile: string,
+		initialData: T,
 		viewColumn: vscode.ViewColumn = vscode.ViewColumn.One,
-		initialData: T
 	) {
-		this._assetUri = vscode.Uri.joinPath(this._context.extensionUri, 'out', 'mssql-react-app');
 		this._panel = vscode.window.createWebviewPanel(
 			'mssql-react-app',
 			title,
@@ -33,7 +30,7 @@ export class ReactWebViewPanelController<T> implements vscode.Disposable {
 			}
 		);
 
-		this._panel.webview.html = this._getHtmlTemplate(this._panel.webview);
+		this._panel.webview.html = this._getHtmlTemplate();
 
 		this._disposables.push(this._panel.webview.onDidReceiveMessage((message) => {
 			if (message.type === 'request') {
@@ -93,10 +90,9 @@ export class ReactWebViewPanelController<T> implements vscode.Disposable {
 		}
 	}
 
-	private _getHtmlTemplate(webView: vscode.Webview) {
+	private _getHtmlTemplate() {
 		const nonce = getNonce();
-		const scriptUri = this.resourceUrl(['assets', this._srcFile]);
-		// const styleUri = this.resourceUrl(['assets', this._styleFile]);
+		const scriptUri = this.resourceUrl([this._srcFile]);
 		return `
 		<!DOCTYPE html>
 				<html lang="en">
@@ -118,11 +114,11 @@ export class ReactWebViewPanelController<T> implements vscode.Disposable {
 				  <script nonce="${nonce}" src="${scriptUri}"></script>
 				</body>
 				</html>
-		`
+		`;
 	}
 
 	public resourceUrl(path: string[]) {
-		return this._panel.webview.asWebviewUri(vscode.Uri.joinPath(this._assetUri, ...path));
+		return this._panel.webview.asWebviewUri(vscode.Uri.joinPath(this._context.extensionUri, 'out', 'react-webviews', 'assets', ...path));
 	}
 
 	public get panel(): vscode.WebviewPanel {
