+++ src/languageservice/decompressProvider.ts
@@ -6,26 +6,16 @@
 import { IDecompressProvider, IPackage } from './interfaces';
 import { ILogger } from '../models/interfaces';
 import * as DecompressTar from 'tar';
-import * as DecompressZip from 'decompress-zip';
-
+import AdmZip from 'adm-zip';
 export default class DecompressProvider implements IDecompressProvider {
 
 	private decompressZip(pkg: IPackage, logger: ILogger): Promise<void> {
-		const unzipper = new DecompressZip(pkg.tmpFile.name);
+		const zip = new AdmZip(pkg.tmpFile.name);
+		logger.appendLine(`Unpacking ${pkg.tmpFile.name} to ${pkg.installPath}...`);
 		return new Promise<void>((resolve, reject) => {
-			let totalFiles = 0;
-			unzipper.on('progress', async (index, fileCount) => {
-				totalFiles = fileCount;
-			});
-			unzipper.on('extract', async () => {
-				logger.appendLine(`Done! ${totalFiles} files unpacked.\n`);
-				resolve();
-			});
-			unzipper.on('error', async (decompressErr) => {
-				logger.appendLine(`[ERROR] ${decompressErr}`);
-				reject(decompressErr);
-			});
-			unzipper.extract({ path: pkg.installPath });
+			zip.extractAllTo(pkg.installPath, true, false);
+			logger.appendLine('Done!\n');
+			resolve();
 		});
 	}
